DIP Project/requirements.txt
streamlit
tensorflow
pillow
numpy
pandas
scikit-learn
